pyproject.toml
[tool.poetry]
name = "herman"
version = "0.1.0"
description = ""
authors = ["leventeharsanyi <hlevente17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
uvicorn = "^0.34.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
jose = "^1.0.0"
python-jose = "^3.3.0"
jinja2 = "^3.1.4"
python-multipart = "^0.0.20"
bcrypt = "^4.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"